requirements.txt
astroid==1.6.5
autopep8==1.4.3
backports.functools-lru-cache==1.5
configparser==3.5.0
enum34==1.1.6
futures==3.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nltk==3.3
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
pycodestyle==2.4.0
pylint==1.9.3
pytesseract==0.2.5
python-dateutil==2.7.5
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
